Functional programming rewrite
