+++ src/room-list/index.ts
@@ -1,59 +1,48 @@
-import { Component, div, select, option } from 'compote';
+import { select, option } from 'compote';
 
 import Room from '../room';
-import SupplyList from '../supply-list';
+import Supply from '../supply';
 
-export default class RoomList extends Component<RoomList> {
-  private rooms: Room[] = this.getRooms();
-  private supplyList: SupplyList;
+import { render } from '../facility';
 
-  render() {
-    return [
-      select({
-        className: 'room-list-select',
-        onchange: ($event: Event) => {
-          const value = (<HTMLSelectElement>$event.target).value;
-          this.rooms.forEach((room) => {
-            if (room.id === value) {
-              this.setSupplyList(room);
-              this.update();
-            }
-          });
-        }
-      }, this.rooms.map((room) => (
-        option({ value: room.id }, room.name))
-      )),
-
-      this.renderSupplyList(this.supplyList)
-    ];
-  }
-
-  private getRooms(): Room[] {
-    const rooms: Room[] = [];
-    const roomsRef = firebase.database().ref('rooms');
-    roomsRef.off('child_added');
-    roomsRef.on('child_added', (roomChildSnapshot: any) => {
-      const room = new Room(roomChildSnapshot.val());
-      room.id = roomChildSnapshot.key;
-      rooms.push(room);
-      if (!this.supplyList) {
-        this.setSupplyList(room);
+export function RoomList(rooms: Room[]) {
+  return (
+    select({
+      className: 'room-list-select',
+      onchange: ($event: Event) => {
+        const value = (<HTMLSelectElement>$event.target).value;
+        rooms.forEach((room) => {
+          if (room.id === value) {
+            getSupplies(rooms, room);
+            render(rooms, room);
+          }
+        });
       }
-      this.update();
+    }, rooms.map((room) => (
+      option({ value: room.id }, room.name))
+    ))
+  );
+}
+
+export function getSupplies(rooms: Room[], room: Room) {
+  const supplies: Supply[] = [];
+  const suppliesRef = firebase.database().ref(`rooms/${room.id}/supplies`);
+  suppliesRef.off('child_added');
+  suppliesRef.on('child_added', (roomSupplyChildSnapshot: any) => {
+    firebase.database().ref(`supplies/${roomSupplyChildSnapshot.key}`).once('value', (supplySnapshot: any) => {
+      const supply = new Supply(supplySnapshot.val());
+      supply.id = supplySnapshot.key;
+      supplies.push(supply);
+      render(rooms, room, supplies);
+
+      // TODO: Unsubscribe on remove
+      firebase.database().ref(`roomSupplies/${room.id}_${supply.id}/requested`).on('value', (requestedSnapshot: any) => {
+        const requested = requestedSnapshot.val();
+        if (requested) {
+          supply.requested = Math.min(requested, Date.now());
+          render(rooms, room, supplies);
+        }
+      });
     });
-
-    return rooms;
-  }
-
-  private setSupplyList(room: Room) {
-    this.supplyList = new SupplyList(this, { room });
-  }
-
-  private renderSupplyList(supplyList: SupplyList): Mithril.VirtualElement {
-    if (!(supplyList && supplyList.room)) return null;
-
-    return div({ className: 'room-list-item' },
-      supplyList.render()
-    );
-  }
+  });
 }
